+++ src/main/java/com/domain/usecase/openfinance/IAccountSynchronizerUseCase.java
@@ -3,6 +3,6 @@ package com.domain.usecase.openfinance;
 import com.domain.entities.AccountEntity;
 
 public interface IAccountSynchronizerUseCase {
-  AccountEntity synchronizeAccount(Integer conceptId, Long cashFlowId, String accountId);
+  AccountEntity synchronizeAccount(Long conceptId, Long cashFlowId, String accountId);
 
 }
